Added a Dynamic Table
