+++ src/components/DynamicTable.tsx
@@ -1,0 +1,357 @@
+import React, { useMemo, Suspense } from "react";
+import {
+  useReactTable,
+  getCoreRowModel,
+  flexRender,
+} from "@tanstack/react-table";
+import type { ColumnDef } from "@tanstack/react-table";
+import FullScreenSpinner from "./FullScreenSpinner";
+
+const SearchIcon = React.lazy(() => import("../assets/svg/Search.svg?react"));
+const FilterIcon = React.lazy(() => import("../assets/svg/Filters.svg?react"));
+const PlusIcon = React.lazy(() => import("../assets/svg/plus.svg?react"));
+
+interface DynamicTableProps<TData extends object> {
+  data: TData[];
+  columns: ColumnDef<TData>[];
+  filterComponent?: React.ReactNode;
+  totalCount: number;
+  currentPage: number;
+  itemsPerPage: number;
+  setCurrentPage: (page: number | ((prev: number) => number)) => void;
+  onFilterStatus?: (status: string) => void;
+  statusFilter?: string;
+  onClearSearch: () => void;
+  onAddNewItem: () => void;
+  isLoading: boolean;
+  error: string | null;
+  title: string;
+  description: string;
+  searchText: string;
+  setSearchText: (text: string) => void;
+  openFilterModal: () => void;
+  applyFilters: () => void;
+  emptyStateMessage: string;
+  emptyStateDescription: string;
+  searchPlaceholder?: string;
+  showStatusFilter?: boolean;
+  filters: object; // Add filters prop
+}
+
+export function DynamicTable<TData extends object>({
+  data,
+  columns,
+  filterComponent,
+  totalCount,
+  currentPage,
+  itemsPerPage,
+  setCurrentPage,
+  onFilterStatus,
+  statusFilter,
+  onClearSearch,
+  onAddNewItem,
+  isLoading,
+  error,
+  title,
+  description,
+  searchText,
+  setSearchText,
+  openFilterModal,
+  applyFilters,
+  emptyStateMessage,
+  emptyStateDescription,
+  searchPlaceholder = "Search",
+  showStatusFilter = true,
+  filters, // Destructure filters prop
+}: DynamicTableProps<TData>) {
+  const table = useReactTable<TData>({
+    data,
+    columns,
+    getCoreRowModel: getCoreRowModel(),
+    getRowId: (originalRow: any, index) =>
+      originalRow.id ? `${originalRow.id}-${index}` : `${index}`,
+  });
+
+  const isFilterActive = useMemo(
+    () => Object.keys(filters ?? {}).length > 0 || searchText.trim() !== "",
+    [filters, searchText]
+  );
+
+  if (error) {
+    return (
+      <div className="w-full h-[75vh] flex items-center justify-center text-red-500 text-lg">
+        {error}
+      </div>
+    );
+  }
+
+  if (isLoading) {
+    return <FullScreenSpinner />;
+  }
+
+  return (
+    <div className="p-6 bg-white shadow-sm dark:bg-[#121418] dark:border-gray-800">
+      <div className="mb-6">
+        <div className="flex items-center justify-between pt-5">
+          <h2 className="text-lg font-semibold text-gray-900 dark:text-white">
+            {title}{" "}
+            <span className="ml-2 text-sm text-blue-600 bg-blue-100 px-2 py-0.5 rounded-full">
+              {totalCount} {title.includes("Rules") ? "Rule" : "Item"}
+              {totalCount !== 1 && "s"}
+            </span>
+          </h2>
+
+          {totalCount !== 0 && (
+            <button
+              onClick={onAddNewItem}
+              className="w-[155px] h-[40px] bg-blue-600 hover:bg-blue-700 text-white px-5 py-2 rounded-md text-sm font-medium"
+            >
+              + Add New {title.includes("Rules") ? "Rule" : "Item"}
+            </button>
+          )}
+        </div>
+
+        <p className="text-sm text-gray-500 mt-1">{description}</p>
+      </div>
+
+      {totalCount === 0 && !isFilterActive ? (
+        <div className="w-full h-[75vh] flex flex-col items-center justify-center bg-gray-50 rounded-md border border-dashed">
+          <div className="bg-white shadow-md rounded-full p-4 mb-4">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-8 w-8 text-gray-500"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+              strokeWidth={2}
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                d="M12 15v2m-6 4h12a2 2 0 002-2v-7a2 2 0 00-2-2h-1V7a5 5 0 00-10 0v3H6a2 2 0 00-2 2v7a2 2 0 002 2z"
+              />
+            </svg>
+          </div>
+          <h3 className="text-lg font-medium text-gray-900 mb-1">
+            {emptyStateMessage}
+          </h3>
+          <p className="text-sm text-gray-500 mb-6">{emptyStateDescription}</p>
+          <button
+            onClick={onAddNewItem}
+            className="bg-blue-600 hover:bg-blue-700 text-white px-5 py-2 rounded-md text-sm font-medium"
+          >
+            + Add New {title.includes("Rules") ? "Rule" : "Item"}
+          </button>
+        </div>
+      ) : (
+        <div className="overflow-x-auto border rounded-lg dark:border-gray-800 ">
+          <div className="flex justify-between items-center px-6 py-6">
+            {showStatusFilter && onFilterStatus && statusFilter && (
+              <div className="flex space-x-0 rounded-lg overflow-hidden border border-gray-300 ">
+                <button
+                  onClick={() => onFilterStatus("All")}
+                  className={`text-xs w-[83px] h-10 px-3 ${
+                    statusFilter === "All"
+                      ? "bg-blue-500 text-white font-semibold"
+                      : "hover:bg-gray-100 text-black dark:hover:bg-gray-200 "
+                  }`}
+                >
+                  View All
+                </button>
+                <button
+                  onClick={() => onFilterStatus("ENABLED")}
+                  className={`text-xs w-[83px] h-10 px-3 border-l ${
+                    statusFilter === "ENABLED"
+                      ? "bg-blue-500 text-white font-semibold"
+                      : "hover:bg-gray-100 text-black dark:text-white dark:hover:bg-gray-800"
+                  }`}
+                >
+                  Active
+                </button>
+                <button
+                  onClick={() => onFilterStatus("DISABLED")}
+                  className={`text-xs w-[83px] h-10 px-3 border-l ${
+                    statusFilter === "DISABLED"
+                      ? "bg-blue-500 text-white font-semibold"
+                      : "hover:bg-gray-100 text-black dark:text-white dark:hover:bg-gray-800"
+                  }`}
+                >
+                  Inactive
+                </button>
+              </div>
+            )}
+
+            {/* Search Box aligned right */}
+            <div className="flex items-center gap-3">
+              <div className="relative w-[400px] h-[44px]">
+                <Suspense>
+                  <SearchIcon className="absolute left-[14px] top-1/2 -translate-y-1/2 text-gray-500 dark:text-gray-400 w-5 h-5 pointer-events-none" />
+                </Suspense>
+                <input
+                  type="text"
+                  value={searchText}
+                  onChange={(e) => setSearchText(e.target.value)}
+                  onKeyDown={(e) => {
+                    if (e.key === "Enter") {
+                      applyFilters();
+                    }
+                  }}
+                  placeholder={searchPlaceholder}
+                  className="w-full h-full pl-[40px] pr-[14px] py-[10px] text-gray-500 rounded-[8px] border border-[#D5D7DA] outline-none focus:ring-1 focus:ring-blue-500 dark:bg-gray-800"
+                />
+              </div>
+
+              {/* Filter Button */}
+              {filterComponent && (
+                <button
+                  className="flex items-center gap-2 w-[100px] h-[40px] px-4 border border-gray-300 rounded-[8px] text-sm text-gray-700 hover:bg-gray-100 dark:text-white dark:hover:bg-gray-800"
+                  onClick={openFilterModal}
+                >
+                  <Suspense>
+                    <FilterIcon className="w-5 h-5 text-gray-500 dark:text-white" />
+                  </Suspense>
+                  Filter
+                </button>
+              )}
+            </div>
+          </div>
+
+          {filterComponent}
+
+          <div
+            className={`w-full ${
+              totalCount === 0 ? "h-[388px] overflow-hidden" : "h-[680px]"
+            } overflow-auto`}
+          >
+            <table className="w-full table-auto h-full text-sm text-center ">
+              <thead className="bg-gray-50 text-gray-600  dark:bg-[#121418] dark:border-gray-800 dark:text-white">
+                {table.getHeaderGroups().map((headerGroup) => (
+                  <tr
+                    key={headerGroup.id}
+                    className="border-b dark:border-gray-800"
+                  >
+                    {headerGroup.headers.map((header) => (
+                      <th key={header.id} className="px-4 h-[72px] font-medium">
+                        {header.isPlaceholder
+                          ? null
+                          : flexRender(
+                              header.column.columnDef.header,
+                              header.getContext()
+                            )}
+                      </th>
+                    ))}
+                  </tr>
+                ))}
+              </thead>
+              <tbody className="w-[1144px] h-[548px]">
+                {table.getRowModel().rows.length > 0 ? (
+                  table.getRowModel().rows.map((row) => (
+                    <tr
+                      key={row.id}
+                      className="border-t hover:bg-gray-50 dark:hover:bg-gray-800 dark:border-gray-800"
+                    >
+                      {row.getVisibleCells().map((cell) => (
+                        <td
+                          key={cell.id}
+                          className="px-4 h-[72px] align-middle"
+                        >
+                          {flexRender(
+                            cell.column.columnDef.cell,
+                            cell.getContext()
+                          )}
+                        </td>
+                      ))}
+                    </tr>
+                  ))
+                ) : (
+                  <tr className="h-full">
+                    <td
+                      colSpan={columns.length}
+                      className="h-full p-0 align-top"
+                    >
+                      <div className="flex items-center justify-center w-[1144px] h-[244px] border-[#E9EAEB]">
+                        <div className="w-[512px] h-[196px] flex items-center justify-center pt-[24px] pb-[24px]">
+                          <div className="w-[352px] h-[196px] gap-[24px]">
+                            {/*Top Section*/}
+                            <div className="w-[352px] h-[132px] flex flex-col items-center gap-[16px]">
+                              {/* Icon */}
+                              <div className="w-[48px] h-[48px] rounded-[28px] border-[8px] border-[#EFF8FF] bg-[#D1E9FF] flex items-center justify-center">
+                                <Suspense>
+                                  <SearchIcon className="text-blue-700" />
+                                </Suspense>
+                              </div>
+                              {/* Text */}
+                              <div className="w-[352px] h-[68px] flex flex-col items-center gap-[4px]">
+                                <h1 className="text-[#181D27] text-[16px] leading-[24px] font-semibold text-center h-[24px]">
+                                  No {title} found
+                                </h1>
+                                <p className="text-[#535862] text-[14px] leading-[20px] text-center h-[40px] pt-[4px]">
+                                  Your search “Keyword” did not match any{" "}
+                                  {title.toLowerCase()}. Please try again or
+                                  create and add a new{" "}
+                                  {title.includes("Rules") ? "rule" : "item"}.
+                                </p>
+                              </div>
+                            </div>
+                            {/*bottom Section*/}
+                            <div className="w-[352px] flex flex-row gap-[12px] pt-[24px] ">
+                              <button
+                                type="button"
+                                onClick={onClearSearch}
+                                className="w-[170px] h-[40px] border border-[#D5D7DA] rounded-[8px] px-[16px] py-[10px] text-[#414651] text-[14px] font-semibold flex items-center justify-center  hover:bg-gray-100"
+                              >
+                                Clear search
+                              </button>
+                              <button
+                                type="button"
+                                className="w-[170px] h-[40px] bg-blue-600 text-white px-[16px] py-[10px] border border-blue-600 rounded-[8px] text-[14px] font-semibold flex items-center justify-center gap-[8px] hover:bg-blue-700"
+                                onClick={onAddNewItem}
+                              >
+                                <Suspense>
+                                  <PlusIcon />
+                                </Suspense>
+                                Add New{" "}
+                                {title.includes("Rules") ? "Rule" : "Item"}
+                              </button>
+                            </div>
+                          </div>
+                        </div>
+                      </div>
+                    </td>
+                  </tr>
+                )}
+              </tbody>
+            </table>
+          </div>
+          {/* Pagination */}
+          <div className="flex justify-between items-center px-4 py-3 border-t dark:border-gray-800 text-sm text-gray-600 w-[1144px] h-[64px]">
+            <div className="pl-6">
+              Page {currentPage} of {Math.ceil(totalCount / itemsPerPage)}
+            </div>
+            <div className="space-x-2 pr-6 text-gray-700">
+              <button
+                onClick={() => setCurrentPage((p) => Math.max(p - 1, 1))}
+                disabled={currentPage === 1}
+                className="px-3 py-2 w-[87px] h-[36px] border border-gray-300 rounded-lg hover:bg-gray-100 disabled:opacity-50 dark:text-white"
+              >
+                Previous
+              </button>
+
+              <button
+                onClick={() =>
+                  setCurrentPage((p) =>
+                    p < Math.ceil(totalCount / itemsPerPage) ? p + 1 : p
+                  )
+                }
+                disabled={currentPage === Math.ceil(totalCount / itemsPerPage)}
+                className="px-3 py-2 w-[60px] h-[36px] border border-gray-300 rounded-lg hover:bg-gray-100 disabled:opacity-50 dark:text-white"
+              >
+                Next
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
